Relocate TaskDetail pages to own directory
