+++ src/js/routes/services.js
@@ -6,10 +6,10 @@ import React from 'react';
 import DeploymentsTab from '../pages/services/DeploymentsTab';
 import ServicesPage from '../pages/ServicesPage';
 import ServicesTab from '../pages/services/ServicesTab';
-import TaskDetail from '../pages/services/task-details/TaskDetail';
-import TaskDetailsTab from '../pages/services/task-details/TaskDetailsTab';
-import TaskFilesTab from '../pages/services/task-details/TaskFilesTab';
-import TaskLogsTab from '../pages/services/task-details/TaskLogsTab';
+import TaskDetail from '../pages/task-details/TaskDetail';
+import TaskDetailsTab from '../pages/task-details/TaskDetailsTab';
+import TaskFilesTab from '../pages/task-details/TaskFilesTab';
+import TaskLogsTab from '../pages/task-details/TaskLogsTab';
 import TaskDetailBreadcrumb from '../pages/nodes/breadcrumbs/TaskDetailBreadcrumb';
 import VolumeDetail from '../components/VolumeDetail';
 import VolumeTable from '../components/VolumeTable';
